# api/proto/v1/blog-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ToDo service",
    "version": "1.0",
    "contact": {
      "name": "go-grpc-http-rest-microservice-tutorial project",
      "url": "https://github.com/amsokol/go-grpc-http-rest-microservice-tutorial",
      "email": "medium@amsokol.com"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/post": {
      "post": {
        "operationId": "CreatePost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreatePostResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreatePostRequest"
            }
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    },
    "/v1/post/{id}": {
      "get": {
        "operationId": "GetPost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetPostResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      },
      "post": {
        "operationId": "CreateComment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateCommentResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateCommentRequest"
            }
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    },
    "/v1/posts": {
      "get": {
        "operationId": "ListPosts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListPostsResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "format": "string"
            }
          }
        },
        "tags": [
          "BlogService"
        ]
      }
    }
  },
  "definitions": {
    "v1Comment": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "v1CreateCommentRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "author": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        }
      }
    },
    "v1CreateCommentResponse": {
      "type": "object",
      "properties": {
        "post": {
          "$ref": "#/definitions/v1Post"
        }
      }
    },
    "v1CreatePostRequest": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "v1CreatePostResponse": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "v1GetPostResponse": {
      "type": "object",
      "properties": {
        "post": {
          "$ref": "#/definitions/v1Post"
        }
      }
    },
    "v1ListPostsResponse": {
      "type": "object",
      "properties": {
        "posts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Post"
          }
        }
      }
    },
    "v1Post": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Comment"
          }
        }
      }
    }
  }
}